% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{data}
\name{hg19.blackList}
\alias{hg19.blackList}
\title{A tibble containing hg19 blackList}
\format{
A tibble with 834 elements
}
\usage{
hg19.blackList
}
\description{
A tibble containing hg19 blackList
}
\examples{
\dontrun{
data("hg19.blackList")
}
}
\keyword{internal}
