% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{data}
\name{hg19.seqLengths}
\alias{hg19.seqLengths}
\title{A array containing hg19 seqlengths}
\format{
A array with 93 elements
}
\usage{
hg19.seqLengths
}
\description{
A array containing hg19 seqlengths
}
\keyword{internal}
