% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{data}
\name{hg38.blackList}
\alias{hg38.blackList}
\title{A tibble containing hg38 blackList}
\format{
A tibble with 635 elements
}
\usage{
hg38.blackList
}
\description{
A tibble containing hg38 blackList
}
\examples{
\dontrun{
data("hg38.blackList")
}
}
\keyword{internal}
