% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{data}
\name{hg38.seqLengths}
\alias{hg38.seqLengths}
\title{A array containing hg38 seqlengths}
\format{
A array with 455 elements
}
\usage{
hg38.seqLengths
}
\description{
A array containing hg38 seqlengths
}
\keyword{internal}
