% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{data}
\name{initial.model}
\alias{initial.model}
\title{A list containing initial models for 2D and 3D}
\format{
A list with 2 elements
}
\usage{
initial.model
}
\description{
A list containing initial models for 2D and 3D
}
\keyword{internal}
