% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{package}
\name{MMSeekR.data}
\alias{MMSeekR.data}
\title{Data for MMSeekR.data}
\description{
MMSeekR.data is data package for PMD callers
\itemize{
  \item initial.model
  \item hg19.blackList
  \item hg19.seqLengths
  \item hg38.blackList
  \item hg38.seqLengths
  \item mm10.blackList
  \item mm10.seqLengths
  \item NNscore.hg19
  \item NNscore.hg38
}
For more information how to create these objects please read the vignette of this package with the
follwing command: \code{browseVignettes("MMSeekR.data")}
}
\examples{
# Please see the datasets
}
\seealso{
\code{\link[MMSeekR.data]{initial.model}}, \code{\link[MMSeekR.data]{hg19.blackList}},
\code{\link[MMSeekR.data]{hg19.seqLengths}}, \code{\link[MMSeekR.data]{hg38.blackList}},
\code{\link[MMSeekR.data]{hg38.seqLengths}}, \code{\link[MMSeekR.data]{NNscore.hg19}},
\code{\link[MMSeekR.data]{NNscore.hg38}}
}
\keyword{utilities}
