% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{data}
\name{NNscore.hg19}
\alias{NNscore.hg19}
\title{A dataframe containing hg19 NNscore}
\format{
A dataframe with 14,647,707 elements
}
\usage{
NNscore.hg19
}
\description{
A dataframe containing hg19 NNscore
}
\examples{
\dontrun{
data("NNscore.hg19")
}
}
\keyword{internal}
