% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSeekR.data.R
\docType{data}
\name{NNscore.hg38}
\alias{NNscore.hg38}
\title{A dataframe containing hg38 NNscore}
\format{
A dataframe with 14,624,755 elements
}
\usage{
NNscore.hg38
}
\description{
A dataframe containing hg38 NNscore
}
\examples{
\dontrun{
data("NNscore.hg38")
}
}
\keyword{internal}
